z test of 2 samples
Suppose the IQ levels among individuals in two different cities are known to be normally distributed with known standard deviations. A researcher wants to know if the mean IQ level between individuals in city A and city B are different, so she selects a simple random sample of  20 individuals from each city and records their IQ levels. The following code shows how to perform a two sample z-test in Python to determine if the mean IQ level is different between the two cities. 
Its pretty simple